UPDATE: chat id fields to big interger

+++ src/entity/user.py
@@ -4,7 +4,7 @@ from entity.base_entity import BaseEntity
 
 
 class User(BaseEntity):
-    id = peewee.IntegerField(primary_key=True)
+    id = peewee.BigIntegerField(primary_key=True)
     username = peewee.CharField(255, null=True)
     first_name = peewee.CharField(255)
     last_name = peewee.CharField(255, null=True)
